requirements.txt
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.31.0
yfinance>=0.2.54
feedparser>=6.0.11
pytest>=8.0.0
